Add a pretty-rat() routine, various updates
Add a new routine to do "pretty printing" of Rats as fractional
values. Folded the documentation into the main library to make
it easier to keep the library and docs in synch. Now generating
the README with Pod-to-Markdown. Added several optionally
imported abbreviated names for the various subs. Not imported
by default; need to specify :short when 'use'ing the module.